Move assets and services registration into the Configurator. (Follow-up from 7fbefe96.)
